// src/gen-ts/src/models/CrupdateLoanRepayment.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Company Management API
 * API CRUD pour la gestion des entreprises, chantiers, entrepôts, équipements, matériaux, mouvements financiers, dépenses, achats et frais de transport. Basée sur le schéma de base de données fourni.
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime'
/**
 *
 * @export
 * @interface CrupdateLoanRepayment
 */
export interface CrupdateLoanRepayment {
  /**
   *
   * @type {string}
   * @memberof CrupdateLoanRepayment
   */
  id?: string
  /**
   *
   * @type {Date}
   * @memberof CrupdateLoanRepayment
   */
  payment_date?: Date
  /**
   * Total repayment amount
   * @type {number}
   * @memberof CrupdateLoanRepayment
   */
  amount?: number
  /**
   * ID of the loan this repayment is related to
   * @type {string}
   * @memberof CrupdateLoanRepayment
   */
  loan_id?: string
  /**
   *
   * @type {string}
   * @memberof CrupdateLoanRepayment
   */
  comment?: string
}

/**
 * Check if a given object implements the CrupdateLoanRepayment interface.
 */
export function instanceOfCrupdateLoanRepayment(value: object): value is CrupdateLoanRepayment {
  return true
}

export function CrupdateLoanRepaymentFromJSON(json: any): CrupdateLoanRepayment {
  return CrupdateLoanRepaymentFromJSONTyped(json, false)
}

export function CrupdateLoanRepaymentFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean,
): CrupdateLoanRepayment {
  if (json == null) {
    return json
  }
  return {
    id: json['id'] == null ? undefined : json['id'],
    payment_date: json['payment_date'] == null ? undefined : new Date(json['payment_date']),
    amount: json['amount'] == null ? undefined : json['amount'],
    loan_id: json['loan_id'] == null ? undefined : json['loan_id'],
    comment: json['comment'] == null ? undefined : json['comment'],
  }
}

export function CrupdateLoanRepaymentToJSON(json: any): CrupdateLoanRepayment {
  return CrupdateLoanRepaymentToJSONTyped(json, false)
}

export function CrupdateLoanRepaymentToJSONTyped(
  value?: CrupdateLoanRepayment | null,
  ignoreDiscriminator: boolean = false,
): any {
  if (value == null) {
    return value
  }

  return {
    id: value['id'],
    payment_date:
      value['payment_date'] == null
        ? value['payment_date']
        : value['payment_date'].toISOString().substring(0, 10),
    amount: value['amount'],
    loan_id: value['loan_id'],
    comment: value['comment'],
  }
}